impl: simplify control flow in SignedShrinker
The condition `(self.x - self.i).abs() < self.x.abs()` is true if
`self.i` is not `0` and has the same sign than `self.x`. The latter is
always the case due to the initialization, the former was recently
introduced to bound the iteration in the case of `self.x == <$ty>::MIN`.
Thus, both of these conditions became redundant.